+++ bot/matches.py
@@ -31,7 +31,15 @@ def addToLobby(player):
     _lobbyList.append(player)
     player.status = PlayerStatus.IS_LOBBIED
     if len(_lobbyList) == cfg.general["lobby_size"]:
+        _autoPing.cancel()
         startMatchFromFullLobby.start()
+    elif len(_lobbyList) >= cfg.general["lobby_size"] - 4:
+        if not _autoPing.is_running():
+            _autoPing.start()
+
+@tasks.loop(seconds=100, delay=1, count=2)
+async def _autoPing():
+    await channelSend("LB_NOTIFY", cfg.discord_ids["lobby"], f'<@&{cfg.discord_ids["notify_role"]}>')
 
 def getLobbyLen():
     return len(_lobbyList)
@@ -41,15 +49,19 @@ def getAllNamesInLobby():
     return names
 
 def removeFromLobby(player):
-    global _lobbyStuck
     _lobbyList.remove(player)
-    _lobbyStuck = False
+    _onLobbyRemove()
     player.status = PlayerStatus.IS_REGISTERED
 
 def _onMatchFree():
     if len(_lobbyList) == cfg.general["lobby_size"]:
         startMatchFromFullLobby.start()
 
+def _onLobbyRemove():
+    global _lobbyStuck
+    _lobbyStuck = False
+    if len(_lobbyList) < cfg.general["lobby_size"]-4:
+        _autoPing.cancel()
 
 @tasks.loop(count=1)
 async def startMatchFromFullLobby():
@@ -57,6 +69,7 @@ async def startMatchFromFullLobby():
     match = _findSpotForMatch()
     if match == None:
         _lobbyStuck = True
+        _autoPing.cancel()
         await channelSend("LB_STUCK", cfg.discord_ids["lobby"])
         return
     _lobbyStuck = False
@@ -86,7 +99,7 @@ def clearLobby():
     for p in _lobbyList:
         p.status = PlayerStatus.IS_REGISTERED
     _lobbyList.clear()
-    _lobbyStuck = False
+    _onLobbyRemove()
     return True
 
 
@@ -264,7 +277,7 @@ class Match():
         await channelSend("MATCH_CONFIRM", self.__id, *captainPings, match=self)
 
     @tasks.loop(minutes=cfg.ROUND_LENGHT, delay=1, count=2)
-    async def __onMatchOver(self):
+    async def _onMatchOver(self):
         playerPings = [tm.allPings for tm in self.__teams]
         await channelSend("MATCH_ROUND_OVER", self.__id, *playerPings, self.roundNo)
         for tm in self.__teams:
@@ -291,7 +304,7 @@ class Match():
         await channelSend("MATCH_STARTED", self.__id, *playerPings, self.roundNo)
         self.__roundsStamps.append(int(dt.timestamp(dt.now())))
         self.__status = MatchStatus.IS_PLAYING
-        self.__onMatchOver.start()
+        self._onMatchOver.start()
 
 
     @tasks.loop(count=1)
@@ -312,7 +325,7 @@ class Match():
         Team and ActivePlayer objects should get garbage collected, nothing is referencing them anymore"""
 
         if self.status == MatchStatus.IS_PLAYING:
-            self.__onMatchOver.cancel()
+            self._onMatchOver.cancel()
             playerPings = [tm.allPings for tm in self.__teams]
             await channelSend("MATCH_ROUND_OVER", self.__id, *playerPings, self.roundNo)
             await channelSend("MATCH_OVER", self.__id)